pyproject.toml
[project]
name = "totalatacadot1"
version = "0.1.0"
description = "Add your description here"
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "briefcase>=0.3.22",
    "loguru>=0.7.3",
    "oracledb>=3.0.0",
    "pyside6>=6.8.2.1",
    "python-dotenv>=1.0.1",
    "qt-material>=2.14",
    "sqlalchemy>=2.0.38",
]

# This project was generated with 0.3.22 using template: https://github.com/beeware/briefcase-template@v0.3.22
[tool.briefcase]
project_name = "Total Atacado T1"
bundle = "com.example"
version = "0.0.1"
url = "https://example.com/totalatacadot1"
license.file = "LICENSE"
author = "Alexandre Lima"
author_email = "alexandrecurylima@gmail.com"

[tool.briefcase.app.totalatacadot1]
formal_name = "Total Atacado T1"
description = "A application for handling integration between total atacado and estapar"
long_description = """More details about the app should go here.
"""
icon = "src/totalatacadot1/assets/icons/icon"
sources = [
    "src/totalatacadot1",
]
test_sources = [
    "tests",
]

requires = [
    "loguru>=0.7.3",
    "pyside6>=6.8.2.1",
    "qt-material>=2.14",
    "oracledb>=3.0.0",
    "sqlalchemy>=2.0.38",
    "python-dotenv>=1.0.1"
]
test_requires = [
    "pytest",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
